Implement keyboard shortcuts (commands).
