+++ commands/commands.html
@@ -1,0 +1,6 @@
+<!DOCTYPE html>
+<html><head>
+<meta charset="utf-8">
+<title>Link Gopher</title>
+<script type="module" src='commands.js'></script>
+</head></html>
